minimal protocol configure
