+++ crates/dwn-test/protocols/minimal.json
@@ -1,0 +1,10 @@
+{
+    "protocol": "http://minimal.xyz",
+    "published": false,
+    "types": {
+        "foo": {}
+    },
+    "structure": {
+        "foo": {}
+    }
+}
